+++ src/main/java/uk/co/reecedunn/intellij/plugin/xquery/parser/XQueryParser.java
@@ -6026,6 +6026,7 @@ class XQueryParser {
                     error(XQueryBundle.message("parser.error.expected", "NCName"));
                     haveError = true;
                 }
+                parseWhiteSpaceAndCommentTokens();
             }
 
             parseWhiteSpaceAndCommentTokens();