+++ CHANGELOG.md
@@ -4,7 +4,10 @@
 
 Bug Fixes:
 
-1.  Fix parsing Saxon 9.8 `tuple` types with occurrence indicators, and multiple tuple fields.
+1.  Fix parsing Saxon 9.8 `tuple` types with whitespace or comments after the
+    item type with an occurrence indicator and before the comma.
+1.  Fix parsing Saxon 9.8 `union` types with whitespace or comments after the
+    QName and before the comma.
 
 Inspections: